mini_App-Langgraph/pyproject.toml
[project]
name = "App-builder-Langgraph"
version = "0.1.0"
description = "LangGraph-based AI app builder that generates complete projects from user prompts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.0.4",
    "langgraph>=0.6.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-version = "3.11"